+++ mpcframework/MPC_Message_Passing/sctk_driver_tcp/sctk_tcp_rdma.c
@@ -84,7 +84,7 @@ static void *sctk_tcp_rdma_thread ( sctk_endpoint_t *tmp )
 
 		res = sctk_safe_read ( fd, ( char * ) &op_type, sizeof ( sctk_tcp_rdma_type_t ) );
 
-		if ( res < sizeof ( sctk_tcp_rdma_type_t ) )
+		if ( res < (ssize_t)sizeof ( sctk_tcp_rdma_type_t ) )
 		{
 			return NULL;
 		}
